+++ src/main.rs
@@ -1,6 +1,6 @@
 use maple_cli::{
     cmd::{Cmd, Maple},
-    subprocess, Result, Source, StructOpt,
+    subprocess, ContentFiltering, Result, Source, StructOpt,
 };
 
 pub mod built_info {
@@ -35,7 +35,7 @@ fn run(maple: Maple) -> Result<()> {
         Cmd::RipgrepForerunner { cmd_dir } => maple_cli::cmd::grep::run_forerunner(
             cmd_dir,
             maple.number,
-            maple.enable_icon,
+            maple.icon_painter,
             maple.no_cache,
         )?,
         Cmd::Cache(cache) => cache.run()?,
@@ -46,6 +46,7 @@ fn run(maple: Maple) -> Result<()> {
             cmd,
             cmd_dir,
             sync,
+            content_filtering,
         } => {
             let source = if let Some(cmd_str) = cmd {
                 if let Some(dir) = cmd_dir {
@@ -64,7 +65,7 @@ fn run(maple: Maple) -> Result<()> {
                     source,
                     algo,
                     maple.number,
-                    maple.enable_icon,
+                    maple.icon_painter,
                     maple.winwidth,
                 )?;
             } else {
@@ -73,28 +74,14 @@ fn run(maple: Maple) -> Result<()> {
                     source,
                     algo,
                     maple.number,
-                    maple.enable_icon,
                     maple.winwidth,
-                    false,
-                    false,
+                    maple.icon_painter,
+                    content_filtering.unwrap_or(ContentFiltering::Full),
                 )?;
             }
         }
-        Cmd::Exec {
-            cmd,
-            output,
-            cmd_dir,
-            output_threshold,
-        } => {
-            maple_cli::cmd::exec::run(
-                cmd,
-                output,
-                output_threshold,
-                cmd_dir,
-                maple.number,
-                maple.enable_icon,
-                maple.no_cache,
-            )?;
+        Cmd::Exec(exec) => {
+            exec.run(maple.number, maple.icon_painter, maple.no_cache)?;
         }
         Cmd::Grep {
             grep_cmd,
@@ -116,7 +103,7 @@ fn run(maple: Maple) -> Result<()> {
                     g,
                     cmd_dir,
                     maple.number,
-                    maple.enable_icon,
+                    maple.icon_painter,
                 )?;
             } else {
                 maple_cli::cmd::grep::dyn_grep(
@@ -124,7 +111,7 @@ fn run(maple: Maple) -> Result<()> {
                     cmd_dir,
                     input,
                     maple.number,
-                    maple.enable_icon,
+                    maple.icon_painter,
                     maple.no_cache,
                 )?;
             }